rmm_refactoring
