+++ CMakePresets.json
@@ -19,7 +19,9 @@
             "inherits": "base",
             "binaryDir": "${sourceDir}/build/debug",
             "cacheVariables": {
-                "CMAKE_BUILD_TYPE": "Debug"
+                "CMAKE_BUILD_TYPE": "Debug",
+                "PIXIE_BENCHMARKS": "OFF",
+                "PIXIE_TESTS": "ON"
             }
         },
         {
@@ -28,7 +30,9 @@
             "inherits": "base",
             "binaryDir": "${sourceDir}/build/release",
             "cacheVariables": {
-                "CMAKE_BUILD_TYPE": "Release"
+                "CMAKE_BUILD_TYPE": "Release",
+                "PIXIE_BENCHMARKS": "OFF",
+                "PIXIE_TESTS": "ON"
             }
         },
         {
@@ -38,7 +42,20 @@
             "binaryDir": "${sourceDir}/build/release",
             "cacheVariables": {
                 "CMAKE_BUILD_TYPE": "Release",
+                "PIXIE_TESTS": "OFF",
                 "PIXIE_BENCHMARKS": "ON"
+            }
+        },
+        {
+            "name": "benchmarks-third-party",
+            "displayName": "Benchmarks with third-party backends",
+            "inherits": "base",
+            "binaryDir": "${sourceDir}/build/release-third-party",
+            "cacheVariables": {
+                "CMAKE_BUILD_TYPE": "Release",
+                "PIXIE_BENCHMARKS": "ON",
+                "PIXIE_TESTS": "OFF",
+                "PIXIE_THIRD_PARTY_BACKENDS": "ON"
             }
         },
         {
@@ -50,6 +67,7 @@
                 "BENCHMARK_ENABLE_LIBPFM": "ON",
                 "CMAKE_BUILD_TYPE": "RelWithDebInfo",
                 "PIXIE_DIAGNOSTICS": "ON",
+                "PIXIE_TESTS": "OFF",
                 "PIXIE_BENCHMARKS": "ON"
             }
         },
@@ -90,17 +108,63 @@
         {
             "name": "debug",
             "displayName": "Build Debug",
-            "configurePreset": "debug"
+            "configurePreset": "debug",
+            "targets": [
+                "unittests",
+                "benchmark_tests",
+                "test_rmm",
+                "louds_tree_tests",
+                "excess_positions_tests"
+            ]
         },
         {
             "name": "release",
             "displayName": "Build Release",
-            "configurePreset": "release"
+            "configurePreset": "release",
+            "targets": [
+                "unittests",
+                "benchmark_tests",
+                "test_rmm",
+                "louds_tree_tests",
+                "excess_positions_tests"
+            ]
+        },
+        {
+            "name": "benchmarks-all",
+            "displayName": "Build Benchmarks",
+            "configurePreset": "benchmarks-all",
+            "targets": [
+                "benchmarks",
+                "bench_rmm",
+                "louds_tree_benchmarks",
+                "alignment_comparison",
+                "excess_positions_benchmarks"
+            ]
+        },
+        {
+            "name": "benchmarks-third-party",
+            "displayName": "Build Benchmarks with third-party backends",
+            "configurePreset": "benchmarks-third-party",
+            "targets": [
+                "benchmarks",
+                "bench_rmm",
+                "bench_rmm_sdsl",
+                "louds_tree_benchmarks",
+                "alignment_comparison",
+                "excess_positions_benchmarks"
+            ]
         },
         {
             "name": "benchmarks-diagnostic",
             "displayName": "Benchmarks diagnostic",
-            "configurePreset": "benchmarks-diagnostic"
+            "configurePreset": "benchmarks-diagnostic",
+            "targets": [
+                "benchmarks",
+                "bench_rmm",
+                "louds_tree_benchmarks",
+                "alignment_comparison",
+                "excess_positions_benchmarks"
+            ]
         },
         {
             "name": "docs",
@@ -113,12 +177,26 @@
         {
             "name": "coverage",
             "displayName": "Build Coverage",
-            "configurePreset": "coverage"
+            "configurePreset": "coverage",
+            "targets": [
+                "unittests",
+                "benchmark_tests",
+                "test_rmm",
+                "louds_tree_tests",
+                "excess_positions_tests"
+            ]
         },
         {
             "name": "asan",
             "displayName": "Build AddressSanitizer",
-            "configurePreset": "asan"
+            "configurePreset": "asan",
+            "targets": [
+                "unittests",
+                "benchmark_tests",
+                "test_rmm",
+                "louds_tree_tests",
+                "excess_positions_tests"
+            ]
         }
     ]
 }
